Increase Rack::Timeout timeout to 8 seconds
This isn't to fix anything, but to help us understand the timeouts that
we've been seeing. Requests are consistenly finishing just *after*
Rack::Timeout raises an exception (i.e. they're always 5.xxx seconds in
the timeout case). If this makes the timeout exceptions go away then
its probably just coincidence, but if the requests suddenly start
timing out but finishing in 8.xxx seconds then something weird is
probably happening with Rack::Timeout and its interaction with this
app.